Split tests for enumerable extensions, fix issues with zero length collections
