+++ Acolyte.NET/Tests/Acolyte.CSharp.Tests/Collections/EnumerableExtensions/FirstOrDefaultTests.cs
@@ -1,0 +1,261 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Acolyte.Common;
+using Acolyte.Tests;
+using Acolyte.Tests.Collections;
+using Acolyte.Tests.Creators;
+
+namespace Acolyte.Collections.Tests.EnumerableExtensions
+{
+    public sealed class FirstOrDefaultTests
+    {
+        public FirstOrDefaultTests()
+        {
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>("source", () => nullValue.FirstOrDefault(default));
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_WithPredicate_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "source", () => nullValue.FirstOrDefault(default, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_WithPredicate_ForNullPredicate()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "predicate", () => emptyCollection.FirstOrDefault(null, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.FirstOrDefault(expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_WithPredicate_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.FirstOrDefault(_ => default, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_FirstOrDefault_ForCollectionWithSomeItems_ShouldReturnFirstItem(int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = collectionWithSomeItems.First();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.FirstOrDefault(default);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_FirstOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldReturnFirstItem(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = collectionWithSomeItems.First();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.FirstOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_FirstOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldReturnDefaultItem(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.FirstOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+            int expectedValue = collectionWithRandomSize.Any()
+                ? collectionWithRandomSize.First()
+                : defaultResult;
+
+            // Act.
+            int actualValue = collectionWithRandomSize.FirstOrDefault(defaultResult);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_WithPredicate_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+            Func<int, bool> predicate = i => TestDataCreator.IsEven(i);
+            int expectedValue = collectionWithRandomSize.Any()
+                ? collectionWithRandomSize.First(predicate)
+                : defaultResult;
+
+            // Act.
+            int actualValue = collectionWithRandomSize.FirstOrDefault(predicate, defaultResult);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_FirstOrDefault_WithPredicate_ForCollectionWithRandomSize_ShouldReturnDefaultItem()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithRandomSize.FirstOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void FirstOrDefault_ShouldLookOnlyAtFirstItemFromCollection()
+        {
+            // Arrange.
+            const int count = 2;
+            var explosiveCollection = ExplosiveCollection.Create(
+                TestDataCreator.CreateRandomInt32List(count),
+                explosiveIndex: Constants.FirstIndex + 1
+            );
+            int expectedValue = explosiveCollection.First();
+
+            // Act.
+            int actualValue = explosiveCollection.FirstOrDefault(default);
+
+            // Assert.
+            Assert.Equal(expected: 1, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void FirstOrDefault_WithPredicate_ShouldStopAfterFoundItemFromCollection()
+        {
+            // Arrange.
+            // Do not use random because we should find exactly second item.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.Create(
+                collection, explosiveIndex: Constants.FirstIndex + 2
+            );
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act.
+            int actualValue = explosiveCollection.FirstOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(expected: 2, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void FirstOrDefault_WithPredicate_ShouldLookWholeCollectionToFindItem()
+        {
+            // Arrange.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.CreateNotExplosive(collection);
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act.
+            int actualValue = explosiveCollection.FirstOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expected: collection.Length, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+    }
+}
