+++ Acolyte.NET/Tests/Acolyte.CSharp.Tests/Collections/EnumerableExtensions/IsNullOrEmptyTests.cs
@@ -1,0 +1,103 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Acolyte.Common;
+using Acolyte.Tests;
+using Acolyte.Tests.Collections;
+using Acolyte.Tests.Creators;
+
+namespace Acolyte.Collections.Tests.EnumerableExtensions
+{
+    public sealed class IsNullOrEmptyTests
+    {
+        public IsNullOrEmptyTests()
+        {
+        }
+
+        [Fact]
+        public void Call_IsNullOrEmpty_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+            Assert.True(nullValue.IsNullOrEmpty());
+        }
+
+        [Fact]
+        public void Call_IsNullOrEmpty_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+
+            // Act & Assert.
+            Assert.True(emptyCollection.IsNullOrEmpty());
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_IsNullOrEmpty_ForCollectionWithSomeItems(int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+
+            // Act & Assert.
+            Assert.False(collectionWithSomeItems.IsNullOrEmpty());
+        }
+
+        [Fact]
+        public void Call_IsNullOrEmpty_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            bool expectedResult = !collectionWithRandomSize.Any();
+
+            // Act.
+            bool actualResult = collectionWithRandomSize.IsNullOrEmpty();
+
+            // Assert.
+            Assert.Equal(expectedResult, actualResult);
+        }
+
+        [Fact]
+        public void Call_IsNullOrEmpty_ForCollectionWithRandomNumberAndNullValues()
+        {
+            // Arrange.
+            IEnumerable<int?> collectionWithRandomSize =
+                TestDataCreator.CreateRandomNullableInt32List();
+            bool expectedResult = !collectionWithRandomSize.Any();
+
+            // Act.
+            bool actualResult = collectionWithRandomSize.IsNullOrEmpty();
+
+            // Assert.
+            Assert.Equal(expectedResult, actualResult);
+        }
+
+        [Fact]
+        public void IsNullOrEmpty_ShouldLookOnlyAtFirstItemFromCollection()
+        {
+            // Arrange.
+            const int count = 2;
+            var explosiveCollection = ExplosiveCollection.Create(
+                TestDataCreator.CreateRandomInt32List(count),
+                explosiveIndex: Constants.FirstIndex + 1
+            );
+            bool expectedResult = !explosiveCollection.Any();
+
+            // Act.
+            bool actualResult = explosiveCollection.IsNullOrEmpty();
+
+            // Assert.
+            Assert.Equal(expected: 1, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedResult, actualResult);
+        }
+    }
+}
