+++ Acolyte.NET/Tests/Acolyte.CSharp.Tests/Collections/EnumerableExtensions/LastOrDefaultTests.cs
@@ -1,0 +1,241 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Acolyte.Tests;
+using Acolyte.Tests.Collections;
+using Acolyte.Tests.Creators;
+
+namespace Acolyte.Collections.Tests.EnumerableExtensions
+{
+    public sealed class LastOrDefaultTests
+    {
+        public LastOrDefaultTests()
+        {
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>("source", () => nullValue.LastOrDefault(default));
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_WithPredicate_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "source", () => nullValue.LastOrDefault(default, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_WithPredicate_ForNullPredicate()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "predicate", () => emptyCollection.LastOrDefault(null, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.LastOrDefault(expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_WithPredicate_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.LastOrDefault(_ => default, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_LastOrDefault_ForCollectionWithSomeItems_ShouldReturnLastItem(int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = collectionWithSomeItems.Last();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.LastOrDefault(default);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_LastOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldReturnLastItem(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = collectionWithSomeItems.Last();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.LastOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_LastOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldReturnDefaultItem(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.LastOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithRandomSize.LastOrDefault(defaultResult);
+
+            // Assert.
+            int expectedValue = collectionWithRandomSize.Any()
+                ? collectionWithRandomSize.Last()
+                : defaultResult;
+
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_WithPredicate_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+            Func<int, bool> predicate = i => TestDataCreator.IsEven(i);
+
+            // Act.
+            int actualValue = collectionWithRandomSize.LastOrDefault(predicate, defaultResult);
+
+            // Assert.
+            int expectedValue = collectionWithRandomSize.Any()
+                ? collectionWithRandomSize.Last(predicate)
+                : defaultResult;
+
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_LastOrDefault_WithPredicate_ForCollectionWithRandomSize_ShouldReturnDefaultItem()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithRandomSize.LastOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void LastOrDefault_WithPredicate_ShouldLookWholeCollectionToFindItemAfterItFoundSomething()
+        {
+            // Arrange.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.CreateNotExplosive(collection);
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act.
+            int actualValue = explosiveCollection.LastOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expected: collection.Length, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void LastOrDefault_WithPredicate_ShouldLookWholeCollectionToFindItem()
+        {
+            // Arrange.
+            // Do not use random because we should find exactly second item.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.CreateNotExplosive(collection);
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act.
+            int actualValue = explosiveCollection.LastOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(collection.Length, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+    }
+}
