+++ Acolyte.NET/Tests/Acolyte.CSharp.Tests/Collections/EnumerableExtensions/SingleOrDefaultTests.cs
@@ -1,0 +1,311 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using Acolyte.Common;
+using Acolyte.Tests;
+using Acolyte.Tests.Collections;
+using Acolyte.Tests.Creators;
+
+namespace Acolyte.Collections.Tests.EnumerableExtensions
+{
+    public sealed class SingleOrDefaultTests
+    {
+        public SingleOrDefaultTests()
+        {
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "source", () => nullValue.SingleOrDefault(default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForNullValue()
+        {
+            // Arrange.
+            const IEnumerable<int>? nullValue = null;
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "source", () => nullValue.SingleOrDefault(default, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForNullPredicate()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+
+            // Act & Assert.
+#pragma warning disable CS8625 // Cannot convert null literal to non-nullable reference type.
+            Assert.Throws<ArgumentNullException>(
+                "predicate", () => emptyCollection.SingleOrDefault(null, default)
+            );
+#pragma warning restore CS8625 // Cannot convert null literal to non-nullable reference type.
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.SingleOrDefault(expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForEmptyCollection()
+        {
+            // Arrange.
+            IEnumerable<int> emptyCollection = Enumerable.Empty<int>();
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = emptyCollection.SingleOrDefault(_ => default, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_ForCollectionWithSingleItem_ShouldReturnFirstItem()
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSingleItem =
+                TestDataCreator.CreateRandomInt32List(TestHelper.OneCollectionSize);
+            int expectedValue = collectionWithSingleItem.Single();
+
+            // Act.
+            int actualValue = collectionWithSingleItem.SingleOrDefault(default);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForCollectionWithSingleItem_ShouldReturnFirstItem()
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSingleItem =
+                TestDataCreator.CreateRandomInt32List(TestHelper.OneCollectionSize);
+            int expectedValue = collectionWithSingleItem.Single();
+
+            // Act.
+            int actualValue = collectionWithSingleItem.SingleOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Theory]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_SingleOrDefault_ForCollectionWithSomeItems_ShouldThrow(int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+
+            // Act & Assert.
+            Assert.Throws(
+                Error.MoreThanOneElement().GetType(),
+                () => collectionWithSomeItems.SingleOrDefault(default)
+            );
+        }
+
+        [Theory]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_SingleOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldThrow(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+
+            // Act & Assert.
+            Assert.Throws(
+                Error.MoreThanOneElement().GetType(),
+                () => collectionWithSomeItems.SingleOrDefault(_ => true, default)
+            );
+        }
+
+        [Theory]
+        [InlineData(TestHelper.OneCollectionSize)]
+        [InlineData(TestHelper.TwoCollectionSize)]
+        [InlineData(TestHelper.FiveCollectionSie)]
+        [InlineData(TestHelper.TenCollectionSize)]
+        [InlineData(TestHelper.HundredCollectionSize)]
+        [InlineData(TestHelper.TenThousandCollectionSize)]
+        public void Call_SingleOrDefault_WithPredicate_ForCollectionWithSomeItems_ShouldReturnDefaultItem(
+            int count)
+        {
+            // Arrange.
+            IEnumerable<int> collectionWithSomeItems = TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithSomeItems.SingleOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IReadOnlyList<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+
+            // Act & Assert.
+            if (collectionWithRandomSize.Count > 1)
+            {
+                Assert.Throws(
+                     Error.MoreThanOneElement().GetType(),
+                     () => collectionWithRandomSize.SingleOrDefault(defaultResult)
+                 );
+            }
+            else
+            {
+                int actualValue = collectionWithRandomSize.SingleOrDefault(defaultResult);
+
+                int expectedValue = collectionWithRandomSize.Any()
+                     ? collectionWithRandomSize.Single()
+                     : defaultResult;
+
+                Assert.Equal(expectedValue, actualValue);
+            }
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForCollectionWithRandomSize()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IReadOnlyList<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int defaultResult = TestDataCreator.CreateRandomInt32();
+            Func<int, bool> predicate = i => TestDataCreator.IsEven(i);
+
+            // Act & Assert.
+            int foundValuesCount = collectionWithRandomSize.Count(predicate);
+            if (foundValuesCount > 1)
+            {
+                Assert.Throws(
+                     Error.MoreThanOneElement().GetType(),
+                     () => collectionWithRandomSize.SingleOrDefault(predicate, defaultResult)
+                 );
+            }
+            else
+            {
+                int actualValue = collectionWithRandomSize.SingleOrDefault(predicate, defaultResult);
+
+                // Collection cannot be empty if we found one value.
+                int expectedValue = foundValuesCount == 1
+                    ? collectionWithRandomSize.Single(predicate)
+                    : defaultResult;
+
+                Assert.Equal(expectedValue, actualValue);
+            }
+        }
+
+        [Fact]
+        public void Call_SingleOrDefault_WithPredicate_ForCollectionWithRandomSize_ShouldReturnDefaultItem()
+        {
+            // Arrange.
+            int count = TestDataCreator.GetRandomCountNumber();
+            IEnumerable<int> collectionWithRandomSize =
+                TestDataCreator.CreateRandomInt32List(count);
+            int expectedValue = TestDataCreator.CreateRandomInt32();
+
+            // Act.
+            int actualValue = collectionWithRandomSize.SingleOrDefault(_ => false, expectedValue);
+
+            // Assert.
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void SingleOrDefault_ShouldLookOnlyAtFirstAndSecondItemsFromCollectionBeforeThrow()
+        {
+            // Arrange.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.Create(
+                collection, explosiveIndex: Constants.FirstIndex + 2
+            );
+            int expectedValue = explosiveCollection.First();
+
+            // Act & Assert.
+            Assert.Throws(
+                Error.MoreThanOneElement().GetType(),
+                () => explosiveCollection.SingleOrDefault(default)
+            );
+
+            Assert.Equal(expected: 2, explosiveCollection.VisitedItemsNumber);
+        }
+
+        [Fact]
+        public void SingleOrDefault_WithPredicate_ShouldLookWholeCollectionToEnsureAppropriateItemIsSingle()
+        {
+            // Arrange.
+            // Do not use random because we should find exactly second item.
+            var collection = new[] { 1, 2, 3, 4 };
+            var explosiveCollection = ExplosiveCollection.CreateNotExplosive(collection);
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act.
+            int actualValue = explosiveCollection.SingleOrDefault(
+                i => i.Equals(expectedValue), default
+            );
+
+            // Assert.
+            Assert.Equal(collection.Length, explosiveCollection.VisitedItemsNumber);
+            Assert.Equal(expectedValue, actualValue);
+        }
+
+        [Fact]
+        public void SingleOrDefault_WithPredicate_ShouldThrowAsSoonAsSecondAppropriateItemWasFound()
+        {
+            // Arrange.
+            // Do not use random because we should find exactly two equal items.
+            var collection = new[] { 1, 2, 3, 2, 5 };
+            var explosiveCollection = ExplosiveCollection.CreateNotExplosive(collection);
+            int expectedValue = explosiveCollection.Skip(1).First();
+
+            // Act & Assert.
+            Assert.Throws(
+             Error.MoreThanOneElement().GetType(),
+             () => explosiveCollection.SingleOrDefault(i => i.Equals(expectedValue), default)
+         );
+
+            Assert.Equal(expected: 4, explosiveCollection.VisitedItemsNumber);
+        }
+    }
+}
